Add a read method to FakeHTTPResponse in tests
There was always supposed to be a read method on this object however
requests used to mask AttributeErrors in this section and so the lack of
the method was never noticed and ignored. With requests commit
327512f5ef9eea79a0cb324ba47e7eabf6967b3a this masking was removed and so
requests-mock tests started failing.

Add a simple read method that returns some response content and fixes
the test.

Change-Id: I2999c9c69b4e9ad895114fab8ae7f8ce275fa2a4
Closes-Bug: #1642697